+++ custom_components/alarmdotcom/hub.py
@@ -26,7 +26,6 @@ class AlarmHub:
 
     def __init__(self, hass: HomeAssistant, config_entry: ConfigEntry) -> None:
         """Initialize the system."""
-
         self.hass: HomeAssistant = hass
         self.config_entry: ConfigEntry = config_entry
 
@@ -42,33 +41,8 @@ class AlarmHub:
 
         self.available: bool = True
 
-    # @property
-    # def available(self) -> bool:
-    #     """
-    #     Whether the Alarm.com API is available.
-
-    #     This will only be true if the websocket connection is established and has not been disconnected
-    #     for more than 60 seconds. This is to prevent the system from being marked as unavailable if the
-    #     connection is temporarily lost.
-    #     """
-    #     # If never connected, treat as unavailable.
-    #     ws = self.api.ws_controller
-    #     if ws.connected:
-    #         # Update last connected time
-    #         self._last_connected = asyncio.get_event_loop().time()
-    #         return True
-
-    #     # If never set, treat as unavailable
-    #     last_connected = getattr(self, "_last_connected", None)
-    #     if last_connected is None:
-    #         return False
-
-    #     # If disconnected for less than 60 seconds, still available
-    #     return bool(asyncio.get_event_loop().time() - last_connected < 60)
-
     async def login(self) -> bool:
         """Log in to alarm.com."""
-
         try:
             await self.api.login()
         except pyadc.AuthenticationFailed as err:
@@ -87,7 +61,6 @@ class AlarmHub:
 
     async def initialize(self) -> bool:
         """Initialize connection to Alarm.com after user-driven authentication has already taken place."""
-
         setup_ok = False
 
         try:
@@ -109,20 +82,19 @@ class AlarmHub:
             if not setup_ok:
                 await self.api.close()
 
-        # Initialize WebSocket connection.
         await self.api.start_event_monitoring(_ws_state_handler)
 
         self.close_jobs.append(self.config_entry.add_update_listener(_update_listener))
 
-        # Create system/hub device.
         device_registry = dr.async_get(self.hass)
 
         device_registry.async_get_or_create(
             config_entry_id=self.config_entry.entry_id,
-            identifiers={(DOMAIN, self.api.active_system.id)},
+            identifiers={(DOMAIN, str(self.api.active_system.id))},
             manufacturer="Alarm.com",
             name=self.api.active_system.name,
             entry_type=dr.DeviceEntryType.SERVICE,
+            model="Security System",
         )
 
         return True
@@ -134,14 +106,14 @@ class AlarmHub:
         Will cancel any scheduled setup retry and will unload
         the config entry.
         """
-
         while self.close_jobs:
             self.close_jobs.pop()()
 
         await self.api.close()
 
         unload_success: bool = await self.hass.config_entries.async_unload_platforms(
-            self.config_entry, PLATFORMS
+            self.config_entry,
+            PLATFORMS,
         )
 
         return unload_success
@@ -149,12 +121,8 @@ class AlarmHub:
 
 async def _ws_state_handler(message: pyadc.EventBrokerMessage) -> None:
     """Handle changes to websocket state for ConfigEntry and logging."""
-
     if not isinstance(message, pyadc.ConnectionEvent):
         return
-
-    # WebSocket service handles reconnection on its own. Handle reporting for DEAD state here; do not attempt to
-    # reconnect independently.
 
     if message.current_state == pyadc.WebSocketState.DEAD:
         log.error("Alarm.com websocket state message: %s", message)
@@ -167,7 +135,6 @@ async def _ws_state_handler(message: pyadc.EventBrokerMessage) -> None:
         log.info("Alarm.com websocket state message: %s", message)
         return
 
-    # Should only print CONNECTED events.
     log.debug("Alarm.com websocket state: %s", message.current_state)